update filter redux
