+++ src/app/components/buttonDrawer/index.js
@@ -1,1 +1,14 @@
-export { default } from './buttonDrawer';
+import { connect } from 'react-redux';
+import buttonDrawer from './buttonDrawer';
+import {
+  clearFilters,
+} from '../../redux/headsets/actions';
+
+const mapStateToProps = () => ({
+});
+
+const mapDispatchToProps = (dispatch) => ({
+  clearFilters: () => dispatch(clearFilters()),
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(buttonDrawer);
