+++ src/app/components/filter/index.js
@@ -1,0 +1,26 @@
+import { connect } from 'react-redux';
+import filter from './filter';
+import {
+  updateFilters,
+  clearFilters,
+} from '../../redux/headsets/actions';
+
+const mapStateToProps = (state) => ({
+  isFiltered: state.headsets.isFiltered,
+  filters: state.headsets.visibilityFilter,
+  compareMode: state.compare.compareMode,
+  outDatated: state.headsets.visibilityFilter.outDatated,
+  comingSoon: state.headsets.visibilityFilter.comingSoon,
+  min: state.headsets.visibilityFilter.priceMin,
+  max: state.headsets.visibilityFilter.priceMax,
+  brands: state.headsets.visibilityFilter.brands,
+  platforms: state.headsets.visibilityFilter.platforms,
+  audience: state.headsets.visibilityFilter.audience,
+});
+
+const mapDispatchToProps = (dispatch) => ({
+  updateFilters: (filters = {}) => dispatch(updateFilters(filters)),
+  clearFilters: () => dispatch(clearFilters()),
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(filter);
